refactor(prosemirror-suggest): rename `Suggestion` to `Suggester`
The name `Suggestion` implies something offered to a user. For example typing `@a` offers a suggestion to tag a certain username. Using the same name to describe the configuration object is confusing.

I'm changing the name back to `Suggester` as the name of the configuration object.

The `Suggester` configures the editor to behave in a desired way so that it can provide suggestions to end users.

- Make `prosemirror-state` and `prosemirror-keymap` peerDependencies.

- Make all `@type/*` peer dependencies optional.